requirements.txt
aiogram==2.23.1
aiosqlite==0.19.0
aiohttp>=3.8.0,<3.9.0
Pillow==10.2.0
pytesseract==0.3.10
